#49 Move handout_standard_test() to test_build_lyx
