+++ source/lib/JMSLab/tests/test_build_lyx.py
@@ -6,6 +6,7 @@ from pathlib import Path
 import unittest
 import shutil
 import os
+from .. import misc
 
 # Import testing helper modules
 from . import _test_helpers as helpers
@@ -13,6 +14,7 @@ from . import _side_effects as fx
 
 from ..builders.build_lyx import build_lyx
 from .._exception_classes import ExecCallError
+
 
 # Define path to the builder for use in patching
 path = 'JMSLab.builders.build_lyx'
@@ -24,6 +26,31 @@ shutil_patch = mock.patch('%s.shutil.copy2' % path)
 TESTDIR = Path(__file__).resolve().parent
 os.chdir(TESTDIR)
 
+
+def handout_standard_test(test_object, builder,
+                  extension   = None,
+                  system_mock = None,
+                  source      = None,
+                  target      = 'test_output.txt',
+                  env         = {}):
+    '''Test that builders run without errors and create logs properly.'''
+    if not source:
+        source = 'input/test_script.%s' % extension
+
+    builder(source = source, target = target, env = env)
+
+    if isinstance(target, str):
+        log_directory = misc.get_directory(target)
+    else:
+        log_directory = misc.get_directory(target[0])
+
+    log_path = os.path.join(log_directory, 'sconscript.log')
+    helpers.check_log(test_object, log_path)
+
+    if system_mock:
+        assert system_mock.call_count == 2
+        system_mock.reset_mock()
+        
 
 class TestBuildLyX(unittest.TestCase):
 
@@ -139,7 +166,7 @@ class TestBuildLyX(unittest.TestCase):
                       'build/path_to_handout__.pdf']
             source = ['input/lyx_test_file.lyx']
             
-            helpers.handout_standard_test(self, build_lyx, 'lyx',
+            handout_standard_test(self, build_lyx, 'lyx',
                                   system_mock = mock_check_output,
                                   target = target,
                                   source = source,
@@ -156,15 +183,17 @@ class TestBuildLyX(unittest.TestCase):
             mock_check_output.side_effect = fx.lyx_side_effect
             
             target = ['build/path_to_clean.pdf', 
-                      'build/path_to_handout_.pdf']
+                      'build/path_to_handout_.pdf',
+                      'build/path_to_handout__.pdf']
             source = ['input/lyx_test_file.lyx']
             
-            helpers.handout_standard_test(self, build_lyx, 'lyx',
+            handout_standard_test(self, build_lyx, 'lyx',
                                   system_mock = mock_check_output,
                                   target = target,
                                   source = source)
             self.assertTrue(os.path.isfile(target[0]))
             self.assertTrue(os.path.isfile(target[1]))
+            self.assertTrue(os.path.isfile(target[2]))
             
             
     def tearDown(self):
